# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Lugom ERP",
        "version": "0.1"
    },
    "paths": {
        "/product/getproductsbyid/{product}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Get Products by ID",
                "description": "Retrieve a list of products by their ID. Returns a list of products associated with the specified store ID.",
                "operationId": "getProductsById",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "store": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Sample Product"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 19.99
                                                    },
                                                    "promotionalPrice": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 15.99
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "https://example.com/image.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/produto/getproductsbystore/{store}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Obter produtos de uma loja específica",
                "description": "Retorna uma lista de produtos de uma loja, filtrando pelo ID da loja.",
                "operationId": "getProductsByStore",
                "parameters": [
                    {
                        "name": "store",
                        "in": "path",
                        "description": "ID da loja",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "store": {
                                            "description": "ID da loja",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID do produto",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "description": "Nome do produto",
                                                        "type": "string",
                                                        "example": "Coca Gelada"
                                                    },
                                                    "price": {
                                                        "description": "Preço do produto",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 179.9
                                                    },
                                                    "promotionalPrice": {
                                                        "description": "Preço promocional do produto",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 161.91
                                                    },
                                                    "image": {
                                                        "description": "URL da imagem do produto",
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "https://example.com/image.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loja ou produtos não encontrados"
                    }
                }
            }
        },
        "/api/produto/getstockbystoreandproduct/{store}/{product}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Obter estoque de um produto específico em uma loja",
                "description": "Retorna o estoque de um produto específico em uma loja específica, filtrando pelos IDs da loja e do produto.",
                "operationId": "getStockByStoreAndProduct",
                "parameters": [
                    {
                        "name": "store",
                        "in": "path",
                        "description": "ID da loja",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estoque do produto retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stock": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID do estoque",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "description": "Quantidade disponível no estoque",
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estoque não encontrado"
                    }
                }
            }
        },
        "/api/produto/getstockbystore/{store}": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Obter estoque de uma loja específica",
                "description": "Retorna uma lista de estoques de todos os produtos disponíveis em uma loja, filtrando pelo ID da loja.",
                "operationId": "getStockByStore",
                "parameters": [
                    {
                        "name": "store",
                        "in": "path",
                        "description": "ID da loja",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de estoques retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stock": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID do estoque",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "amount": {
                                                        "description": "Quantidade disponível no estoque",
                                                        "type": "integer",
                                                        "example": 10
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Loja ou estoques não encontrados"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "product",
            "description": "product"
        }
    ]
}